Modified Modules Folder Structure
