+++ src/Modules.js
@@ -2,11 +2,11 @@
  * Core modules
  */
 module.exports = {
-  'do-nothing': require('./modules/DoNothing'),
-  'green-channel': require('./modules/GreenChannel'),
-  'ndvi-red': require('./modules/NdviRed'),
-  'do-nothing-pix': require('./modules/DoNothingPix'),
-  'invert': require('./modules/Invert'),
-  'crop': require('./modules/Crop'),
-  'segmented-colormap': require('./modules/SegmentedColormap')
+  'do-nothing': require('./modules/DoNothing/Module'),
+  'green-channel': require('./modules/GreenChannel/Module'),
+  'ndvi-red': require('./modules/NdviRed/Module'),
+  'do-nothing-pix': require('./modules/DoNothingPix/Module.js'),
+  'invert': require('./modules/Invert/Module'),
+  'crop': require('./modules/Crop/Module'),
+  'segmented-colormap': require('./modules/SegmentedColormap/Module')
 }
