+++ src/modules/Threshold/ImageThreshold.js
