+++ src/modules/_nomodule/PixelManipulation.js
